# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Practica de laravel",
        "version": "2.0"
    },
    "paths": {
...
        "/api/students": {
            "get": {
                "tags": [
                    "Estudiante"
                ],
                "summary": "Obtener lista de estudiantes",
                "operationId": "4618e29c551d12a832be0c4aead5f47c",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "lastName",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "phone",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una lista de estudiantes"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/student/{id}": {
            "get": {
                "tags": [
                    "Estudiante"
                ],
                "summary": "Obtener un estudiante",
                "operationId": "834b7fb95f13081e492cf084dc9cfa19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "put": {
                "tags": [
                    "Estudiante"
                ],
                "summary": "actualizar un estudiante",
                "operationId": "2871729e5d84df1c1bc33c536bbf337c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Datos necesarios para actualizar un estudiante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName",
                                    "email",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Laptop"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": 1
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "float",
                                        "example": 1200.5
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "Descripción opcional del producto"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Descripción opcional del producto"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Una estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Estudiante"
                ],
                "summary": "eliminar un estudiante",
                "operationId": "7d15c12d59c774e71b8c917637c7a6a2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Una estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
        "/api/student": {
            "post": {
                "tags": [
                    "Estudiante"
                ],
                "summary": "crear un estudiante",
                "operationId": "b47d1928ef06188e14585da40c9e8245",
                "requestBody": {
                    "description": "Datos necesarios para crear un estudiante",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "lastName",
                                    "email",
                                    "phone",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Clever"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "SantaMar"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "float",
                                        "example": "Test@gmail.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "809-923-1234"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Algun lugar del planeta"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crea un estudiante"
                    },
                    "default": {
                        "description": "Ha ocurrido un error"
                    }
                }
            }
        },
...
    },
...
}